Split temperature controls into own files and give generic names
